refactor(web): remove account pool ts-nocheck debt
Extract typed UpstreamAccounts helpers and restore strict typing across account-pool production modules.
