+++ docs/specs/README.md
@@ -162,3 +162,4 @@
 | mpgea | Dashboard 活动总览增加“昨日”页签 | 已完成 | `mpgea-dashboard-yesterday-activity-overview/SPEC.md` | 2026-04-11 | fast-track / follow-up of `#r99mz` + `#2qsev` / add fixed previous-natural-day dashboard tab with today-style KPI + minute chart / summary+timeseries accept `yesterday` / targeted cargo+vitest+build+storybook green / Storybook evidence archived in spec / review-loop clear / merge-ready |
 | huzqt | 前端结构收敛 follow-up | 已完成（PR #331） | `huzqt-frontend-structure-convergence-followup/SPEC.md` | 2026-04-11 | fast-track / account-pool page decomposition + typed follow-up / storybook evidence shown / shared-testbox browser smoke green / merged + cleanup |
 | phb37 | 后端结构收敛 follow-up | 已完成 | `phb37-backend-structure-convergence-followup/SPEC.md` | 2026-04-11 | fast-track / maintenance archive + hourly-rollup helpers converged into real support modules / local cargo fmt+check+test green / shared-testbox proxy+raw smoke green |
+| ts6qp | account-pool 前端类型债清偿 | 已完成 | `ts6qp-account-pool-typing-debt/SPEC.md` | 2026-04-12 | normal-flow / production `@ts-nocheck` removed from account-pool pages / `UpstreamAccounts` helper extraction + Create strict typing landed / local lint+test+build+storybook green / visual evidence archived |
